agent_http/requirements.txt
google-adk
mcp
google-auth
google-auth-httplib2
requests
python-dotenv
